refactor: 공통 fetch 함수 분리
